requirements.txt
Flask~=2.0.1
flask_sqlalchemy
flask_wtf
flask_login
email_validator
Pillow
